Move save&restore class to common package
