+++ app/save-and-restore/model/src/main/java/org/phoebus/applications/saveandrestore/model/RestoreResult.java
@@ -1,8 +1,6 @@
-package org.phoebus.service.saveandrestore.epics;
+package org.phoebus.applications.saveandrestore.model;
 
 import java.io.Serializable;
-
-import org.phoebus.applications.saveandrestore.model.SnapshotItem;
 
 public class RestoreResult implements Serializable {
     private SnapshotItem snapshotItem;
